requirements.txt
Flask==2.3.2
gunicorn==20.1.0
google-cloud-storage==2.10.0
sendgrid==6.9.7
requests==2.31.0
python-dotenv==1.0.0
